gtp6.d/e

+++ src/plugins/srv6-mobile/mobile.h
@@ -31,7 +31,7 @@
 typedef struct srv6_end_gtp6_param_s
 {
   ip6_address_t sr_prefix;
-  u16 sr_prefixlen;
+  u32 sr_prefixlen;
 } srv6_end_gtp6_param_t;
 
 typedef struct srv6_end_main_v4_s
